Fix convert_engine under Windows
According to
https://docs.python.org/2/library/tempfile.html#tempfile.NamedTemporaryFile
it is not possible under Windows that a temporary file, which is opened
by the python process via NamedTemporaryFile, is used by another process
at the same time. Therefore starting the subprocess inside the with
NamedTemporaryFile block fails because the started process can not
access the created temporary file. The usage of tempfile.mkstemp allows
that the temporary file is not opened when the subprocess is started.