Use tempfile.mkstemp instead of NamedTemporaryFile
With Python stdlib tempfile.NamedTemporaryFile or Django
django.core.files.temp.NamedTemporaryFile it is not possible to open the
temporary file in another process from within a with
NamedTemporaryFile(...) as fp: block. Also Django's implementation lacks
the support for several arguments like delete on Windows. This
combination of characteristics of the two NamedTemporaryFile
implementations makes the convert engine unusable under Windows at the
moment. By using tempfile.mkstemp it is possible to open the file in
another subprocess even under Windows and the issues are resolved.