Update README inline with docstring on ImageField
The README currently implies that the ImageField will automatically delete references to itself in the key
value store and its thumbnail references and the thumbnail files when deleted. However, the inline docstrings on the ImageField itself, says that this is only fired in Django 1.2.5 or less. This change updates the README to be inline with that information.